add error handling for numbers and empty input
